requirement.txt
Braindecode==0.4.85
captum==0.7.0
certifi==2024.2.2
charset-normalizer3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
docstring-inheritance2.2.0
einops==0.7.0
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.5.0
h5py==3.11.0
idna==3.7
importlib_metadata7.1.0
importlib_resources6.4.0
intel-openmp2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl==2021.4.0
mne==1.6.1
mpmath==1.3.0
networkx==3.0
numba==0.58.1
numpy==1.24.1
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pip==24.0
platformdirs4.2.2
pooch==1.8.1
protobuf==5.26.0
pyparsing==3.1.2
python-dateutil2.9.0.post0
pytz==2024.1
requests==2.31.0
resampy==0.4.3
scikit-learn1.3.2
scipy==1.10.1
setuptools==69.5.1
six==1.16.0
skorch==0.15.0
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tensorboardX2.6.2.2
threadpoolctl3.5.0
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchinfo==1.8.0
torchsummary1.5.1
torchvision==0.18.0+cu118
tqdm==4.66.4
typing_extensions4.9.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
xlwt==1.3.0
zipp==3.18.2
